+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/AnilistModels.kt
@@ -56,7 +56,7 @@ data class ALUserManga(
         score = score_raw.toFloat()
         started_reading_date = start_date_fuzzy
         finished_reading_date = completed_date_fuzzy
-        last_chapter_read = chapters_read
+        last_chapter_read = chapters_read.toFloat()
         library_id = this@ALUserManga.library_id
         total_chapters = manga.total_chapters
     }
